Consolidate envio TypeScript surface into a single index.d.ts
Folds `src/Types.ts` into `index.d.ts` and drops now-unused glue that
only existed for the genType bridge:
- generic block types (`BlockEvent`, `FuelBlockEvent`, `SvmOnBlockArgs`,
  `OnBlockArgs`, `OnBlockOptions`) — superseded by the project-scoped
  `Evm*`/`Fuel*`/`Svm*` variants,
- the `Utils.bigIntSchema` re-export alias (index.js imports via
  `$$BigInt.schema` directly now),
- the top-level `type id = string` in generated `Indexer.res` (entity
  modules already declare their own),
- the six `.res.d.mts` shims that were silencing tsc on genType output,
- a stale `*.res.js` gitignore entry (ReScript 12 emits `.res.mjs`).

`scenarios/test_codegen` handler updated to use `string` instead of
`Indexer.id`.
